third_party/WebKit/LayoutTests/fast/multicol/dynamic/change-second-row-height.html

<!DOCTYPE html>
<p>The word "HEST" should be seen below.</p>
<div style="position:relative; columns:2; column-gap:0; column-fill:auto; width:400px; height:40px; line-height:20px; orphans:1; widows:1;">
    <div id="inner" style="columns:2; column-gap:0; column-fill:auto; height:80px;">
        <div>
            <br>
            <br>
            <br>
            <br>
            <br>
            <span id="elm">HEST</span><br>
        </div>
    </div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(() => {
    var inner = document.getElementById("inner");
    var elm = document.getElementById("elm");
    assert_equals(elm.offsetLeft, 200);
    document.body.offsetLeft;
    inner.style.height = "72px";
    assert_equals(elm.offsetLeft, 300);
}, "Change inner multicol height, which only affects the second row.");
</script>
